requirements.txt
aiohttp
asyncio
tqdm
tenacity
pyyaml